Refactoring of layer class structure
