+++ lib/util/helper.dart
@@ -22,7 +22,11 @@ import 'dart:ui' as ui;
 import 'package:flutter/foundation.dart';
 import 'package:kpix/layer_states/drawing_layer_state.dart';
 import 'package:kpix/layer_states/layer_state.dart';
-import 'package:kpix/managers/history_manager.dart';
+import 'package:kpix/managers/history/history_color_reference.dart';
+import 'package:kpix/managers/history/history_drawing_layer.dart';
+import 'package:kpix/managers/history/history_manager.dart';
+import 'package:kpix/managers/history/history_ramp_data.dart';
+import 'package:kpix/managers/history/history_state.dart';
 import 'package:kpix/util/file_handler.dart';
 import 'package:kpix/util/typedefs.dart';
 import 'package:path/path.dart' as p;
@@ -653,6 +657,20 @@ class Helper
     }
 
     return indexMap;
+  }
+
+  static int? getRampIndex({required String uuid, required final List<HistoryRampData> ramps})
+  {
+    int? rampIndex;
+    for (int i = 0; i < ramps.length; i++)
+    {
+      if (ramps[i].uuid == uuid)
+      {
+        rampIndex = i;
+        break;
+      }
+    }
+    return rampIndex;
   }
 
 }
